+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+#### **6.9**
+
+- Added the possibility to move tabs from the Global section to the Profile section and vice versa
+
+#### **6.8**
+
+- Added the import/export feature (v1.0). This means you can now easily share any of your tabs, through the use of an automatically generated text
+
+#### **6.7**
+
+- Added support for WoW Classic versions
+- Added localization for Korean (koKR)
+
 #### **6.6**
 
 - Global tabs are here! They are account-wide and work just like normal tabs. You can create them in the addon options